splash screen fixed and some warnings fixed

+++ Pic6/NSMutableArray+Shuffle.m
@@ -9,7 +9,7 @@
     NSUInteger count = [self count];
     for (NSUInteger i = 0; i < count; ++i) {
         NSInteger remainingCount = count - i;
-        NSInteger exchangeIndex = i + arc4random_uniform(remainingCount);
+        NSInteger exchangeIndex = i + arc4random_uniform((unsigned int)remainingCount);
         [self exchangeObjectAtIndex:i withObjectAtIndex:exchangeIndex];
     }
 }
